feat: create types module and update components naming
