+++ src/common/types.ts
@@ -1,0 +1,9 @@
+export interface IProduct {
+  id: string
+  title: string
+  price: number
+  rating: number
+  category: string
+  description: string
+  image: string
+}
